[MPT] Code refactoring, added Node file and package object
