+++ src/main/scala/io/iohk/ethereum/merklepatriciatrie/MerklePatriciaTrie.scala
@@ -477,78 +477,6 @@ class MerklePatriciaTrie[K, V](private val rootHash: Option[Array[Byte]],
   }
 }
 
-trait ByteArraySerializable[T] {
-  def toBytes(input: T): Array[Byte]
-
-  def fromBytes(bytes: Array[Byte]): T
-}
-
-trait DataSource {
-
-  type Key = Array[Byte]
-  type Value = Array[Byte]
-
-  def get(key: Key): Option[Value]
-
-  def update(rootHash: Array[Byte], toRemove: Seq[Key], toUpdate: Seq[(Key, Value)]): DataSource
-}
-
-/**
-  * Trie elements
-  */
-private sealed trait Node {
-
-  import MerklePatriciaTrie._
-
-  private var hashedCache: Option[Array[Byte]] = None
-
-  lazy val encode: Array[Byte] = encodeRLP[Node](this)
-
-  def hash(hashFn: HashFn): Array[Byte] = hashedCache.getOrElse {
-    val hashed = hashFn(encode)
-    hashedCache = Some(hashed)
-    hashed
-  }
-
-  def capped(hashFn: HashFn): Array[Byte] = {
-    val encoded = encode
-    if (encoded.length < 32) encoded else hash(hashFn)
-  }
-}
-
-private case class LeafNode(key: Array[Byte], value: Array[Byte]) extends Node
-
-private case class ExtensionNode(sharedKey: Array[Byte], next: Either[Array[Byte], Node]) extends Node
-
-private object ExtensionNode {
-  def apply(sharedKey: Array[Byte], next: Node, hashFn: HashFn): ExtensionNode = {
-    val nextCapped = next.capped(hashFn)
-    ExtensionNode(sharedKey, if (nextCapped.length == 32) Left(nextCapped) else Right(next))
-  }
-}
-
-private case class BranchNode(children: Seq[Option[Either[Array[Byte], Node]]], terminator: Option[Array[Byte]]) extends Node {
-  def updateChild(childIndex: Int, childNode: Node, hashFn: HashFn): BranchNode = {
-    val childCapped = childNode.capped(hashFn)
-    BranchNode(children.updated(childIndex, Some(if (childCapped.length == 32) Left(childCapped) else Right(childNode))), terminator)
-  }
-}
-
-private object BranchNode {
-  private val emptyChildren: Seq[Option[Either[Array[Byte], Node]]] = Array.fill(MerklePatriciaTrie.ListSize - 1)(None)
-
-  def withValueOnly(terminator: Array[Byte]): BranchNode =
-    BranchNode(emptyChildren, Some(terminator))
-
-  def withSingleChild(position: Byte, child: Node, terminator: Option[Array[Byte]], hashFn: HashFn): BranchNode = {
-    val childCapped = child.capped(hashFn)
-    BranchNode(emptyChildren.updated(position, Some(if (childCapped.length == 32) Left(childCapped) else Right(child))), terminator)
-  }
-
-  def withSingleChild(position: Byte, child: Either[Array[Byte], Node], terminator: Option[Array[Byte]]): BranchNode =
-    BranchNode(emptyChildren.updated(position, Some(child)), terminator)
-}
-
 private case class NodeInsertResult(newNode: Node,
                                     toDeleteFromStorage: Seq[Node] = Seq(),
                                     toUpdateInStorage: Seq[Node] = Seq())
